+++ test/pebble.h
@@ -233,6 +233,7 @@ extern uint32_t MESSAGE_KEY_WEATHER_HIGH;
 extern uint32_t MESSAGE_KEY_WEATHER_LOW;
 extern uint32_t MESSAGE_KEY_WEATHER_AQI;
 extern uint32_t MESSAGE_KEY_WEATHER_UV;
+extern uint32_t MESSAGE_KEY_WEATHER_UV_NOW;
 extern uint32_t MESSAGE_KEY_SLOT_1;
 extern uint32_t MESSAGE_KEY_SLOT_2;
 extern uint32_t MESSAGE_KEY_SLOT_3;